drop support for ogranization settings
