+++ buildkite/data_source_organization_test.go
@@ -1,40 +1,37 @@
 package buildkite
 
 import (
+	"encoding/json"
 	"fmt"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
 func TestAccDataOrganization(t *testing.T) {
-	t.Parallel()
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: protoV6ProviderFactories(),
-		CheckDestroy:             testCheckOrganizationSettingsResourceRemoved,
-		Steps: []resource.TestStep{
-			{
-				Config: testDatasourceOrganization(),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					// Confirm that the allowed IP addresses are set correctly in Buildkite's system
-					testAccCheckOrganizationSettingsRemoteValues([]string{"0.0.0.0/0", "1.1.1.1/32", "1.0.0.1/32"}),
-					// Check that the second IP added to the list is the one we expect, this also ensures the length is greater than 1
-					// allowing us to assert the first IP is also added correctly
-					resource.TestCheckResourceAttr("data.buildkite_organization.settings", "allowed_api_ip_addresses.2", "1.0.0.1/32"),
-				),
+	t.Run("organization data source can be loaded from slug", func(t *testing.T) {
+		resource.ParallelTest(t, resource.TestCase{
+			PreCheck:                 func() { testAccPreCheck(t) },
+			ProtoV6ProviderFactories: protoV6ProviderFactories(),
+			Steps: []resource.TestStep{
+				{
+					Config: testAccOrganizationSettingsConfigBasic([]string{"0.0.0.0/0", "1.1.1.1/32", "1.0.0.1/32"}),
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr("data.buildkite_organization.settings", "allowed_api_ip_addresses.2", "1.0.0.1/32"),
+					),
+				},
 			},
-		},
+		})
 	})
+
 }
 
-func testDatasourceOrganization() string {
-	data := `
-	%s
-	data "buildkite_organization" "settings" {
-	  depends_on = [buildkite_organization_settings.let_them_in]
-	}
+func testAccOrganizationSettingsConfigBasic(ip_addresses []string) string {
+	config := `
+		data "buildkite_organization" "settings" {
+      		allowed_api_ip_addresses = %v
+		}
 	`
-	return fmt.Sprintf(data, testAccOrganizationSettingsConfigBasic([]string{"0.0.0.0/0", "1.1.1.1/32", "1.0.0.1/32"}))
+	marshal, _ := json.Marshal(ip_addresses)
+	return fmt.Sprintf(config, string(marshal))
 }
