+++ buildkite/provider.go
@@ -148,79 +148,3 @@ func New(version string) provider.Provider {
 		version: version,
 	}
 }
-
-// Provider creates the schema.Provider for Buildkite
-func Provider(version string) *schema.Provider {
-	provider := &schema.Provider{
-		ResourcesMap: map[string]*schema.Resource{
-			"buildkite_organization_settings": resourceOrganizationSettings(),
-		},
-		Schema: map[string]*schema.Schema{
-			SchemaKeyOrganization: {
-				Description: "The Buildkite organization slug",
-				Optional:    true,
-				Type:        schema.TypeString,
-			},
-			SchemaKeyAPIToken: {
-				Description: "API token with GraphQL access and `write_pipelines, read_pipelines` and `write_suites` REST API scopes",
-				Optional:    true,
-				Type:        schema.TypeString,
-				Sensitive:   true,
-			},
-			SchemaKeyGraphqlURL: {
-				Description: "Base URL for the GraphQL API to use",
-				Optional:    true,
-				Type:        schema.TypeString,
-			},
-			SchemaKeyRestURL: {
-				Description: "Base URL for the REST API to use",
-				Optional:    true,
-				Type:        schema.TypeString,
-			},
-			"archive_pipeline_on_delete": {
-				Description: "Archive pipelines when destroying instead of completely deleting.",
-				Optional:    true,
-				Type:        schema.TypeBool,
-			},
-		},
-	}
-	provider.ConfigureFunc = providerConfigure(provider.UserAgent("buildkite", version))
-
-	return provider
-}
-
-func providerConfigure(userAgent string) func(d *schema.ResourceData) (interface{}, error) {
-	return func(d *schema.ResourceData) (interface{}, error) {
-		apiToken := os.Getenv("BUILDKITE_API_TOKEN")
-		graphqlUrl := defaultGraphqlEndpoint
-		organization := getenv("BUILDKITE_ORGANIZATION_SLUG")
-		restUrl := defaultRestEndpoint
-
-		if v, ok := d.Get(SchemaKeyAPIToken).(string); ok && v != "" {
-			apiToken = v
-		}
-		if v, ok := d.Get(SchemaKeyGraphqlURL).(string); ok && v != "" {
-			graphqlUrl = v
-		} else if v, ok := os.LookupEnv("BUILDKITE_GRAPHQL_URL"); ok {
-			graphqlUrl = v
-		}
-		if v, ok := d.Get(SchemaKeyOrganization).(string); ok && v != "" {
-			organization = v
-		}
-		if v, ok := d.Get(SchemaKeyRestURL).(string); ok && v != "" {
-			restUrl = v
-		} else if v, ok := os.LookupEnv("BUILDKITE_REST_URL"); ok {
-			restUrl = v
-		}
-
-		config := &clientConfig{
-			org:        organization,
-			apiToken:   apiToken,
-			graphqlURL: graphqlUrl,
-			restURL:    restUrl,
-			userAgent:  userAgent,
-		}
-
-		return NewClient(config)
-	}
-}
